#56477 게시판 프로젝트 수정 삭제 권한 추가
•url로 작성자가 아닌 사용자가 modify나 delete로 직접 타이핑하여 수정 삭제 할 수 없도록 spring security를 이용하여 권한 추가

http://iscr01.navercorp.com/issues/56477

+++ src/main/java/com/naver/jihyunboard/board/controller/BoardController.java
@@ -88,13 +88,12 @@ public class BoardController {
 		HttpServletResponse response, Authentication auth,
 		@RequestParam("currentPage") int currentPage,
 		@RequestParam("searchOption") String searchOption, @RequestParam("keyword") String keyword) throws Exception {
-		//작성자가 아니라면 에러처리
-
-		/*auth = SecurityContextHolder.getContext().getAuthentication();
+		auth = SecurityContextHolder.getContext().getAuthentication();
 		String userId = auth.getName();
-		if (userId != boardService.writerId(boardNum)) {
+
+		if (!userId.equals(boardService.writerId(boardNum))) {
 			return "/board/authError";
-		}*/
+		}
 
 		List<String> list = boardService.getFileList(boardNum);
 		model.addAttribute("list", list);
@@ -102,6 +101,7 @@ public class BoardController {
 		model.addAttribute("currentPage", currentPage);
 		model.addAttribute("keyword", keyword);
 		model.addAttribute("searchOption", searchOption);
+
 		return "/board/modify";
 	}
 
@@ -118,7 +118,13 @@ public class BoardController {
 	}
 
 	@RequestMapping(value = "/delete")
-	public String delete(@RequestParam("boardNum") int boardNum) throws Exception {
+	public String delete(@RequestParam("boardNum") int boardNum, Authentication auth) throws Exception {
+		auth = SecurityContextHolder.getContext().getAuthentication();
+		String userId = auth.getName();
+		if (!userId.equals(boardService.writerId(boardNum))) {
+			return "/board/authError";
+		}
+
 		boardService.deleteBoard(boardNum);
 		return "redirect:list";
 	}
